acmicpc 2530 수정

+++ src/main/kotlin/october/acmicpc2530.kt
@@ -4,8 +4,8 @@ import java.util.*
 import kotlin.math.abs
 
 fun main(args: Array<String>) = with(Scanner(System.`in`)) {
-    val n = nextInt()
-    val m = nextInt()
+    val n = nextLong()
+    val m = nextLong()
     val minusValue = n - m
     val answer = abs(minusValue)
